+++ uSEQ/src/dsp/MedianFilter.h
@@ -46,14 +46,14 @@ public:
 
     double std() {
         double sum = std::accumulate(std::begin(circularBuffer_), std::end(circularBuffer_), 0.0);
-        double m =  sum / circularBuffer_.size();
+        double m =  sum / static_cast<double>(circularBuffer_.size());
 
         double accum = 0.0;
         std::for_each (std::begin(circularBuffer_), std::end(circularBuffer_), [&](const double d) {
             accum += (d - m) * (d - m);
         });
 
-        return sqrt(accum / (circularBuffer_.size()-1));
+        return sqrt(accum / static_cast<double>(circularBuffer_.size()-1));
     }
 
 private:
